move job data models to job data file
