--- internal/models/scan_path_data.go
@@ -1,7 +1,0 @@
-package models
-
-import "github.com/google/uuid"
-
-type ScanPathData struct {
-	LibraryPathId uuid.UUID `json:"libraryPathId"`
-}
